further abstracted the fileLocations into seperate Trait
